Move Turso migration logic into alembic/env.py
Now `alembic upgrade head` works everywhere:
- Local: uses SQLite via SQLAlchemy
- Render: uses Turso via HTTP API (auto-detected from env vars)

No separate script needed.
